Convert image to object url for preview

+++ src/app/api/createlisting/route.js
@@ -1,29 +1,44 @@
+"use server";
 import { writeFile } from "fs/promises";
+import fs from "fs";
+
 export async function POST(req, res) {
   const formData = await req.formData();
-  console.log(formData.get("image"));
+
   const listingImage = formData.get("image");
-  console.log(listingImage.type);
   const contentType = listingImage.type;
   const mimeType = contentType.split("/")[1];
   const listingImageBuffer = await listingImage.arrayBuffer();
 
-  await writeFile(
-    `image.${mimeType}`,
-    Buffer.from(listingImageBuffer),
-    (err) => {
-      if (err) {
-        console.error("Error saving image: ", err);
-      } else {
-        console.log("Image saved successfully");
-      }
-    },
-  );
+  // will need to check if listingTitle already exists
+  const listingTitle = formData.get("title");
+  const listingPrice = formData.get("price");
+  const listingDescription = formData.get("description");
 
-  //try {
-  //  await writeFile("image.png", base64ListingImage);
-  //} catch (e) {
-  //  console.error(e);
-  //}
-  return new Response(200);
+  let response = new Response(500);
+  // creates directory for listing if it doesn't exist
+  if (!fs.existsSync(`public/${listingTitle}`)) {
+    fs.mkdirSync(`public/${listingTitle}`);
+    await writeFile(
+      `public/${listingTitle}/listing-image.${mimeType}`,
+      Buffer.from(listingImageBuffer),
+      (err) => {
+        if (err) {
+          console.error("Error saving image: ", err);
+        } else {
+          console.log("Image saved successfully");
+        }
+      },
+    );
+    await writeFile(`public/${listingTitle}/price.txt`, listingPrice);
+    await writeFile(
+      `public/${listingTitle}/description.txt`,
+      listingDescription,
+    );
+    response = new Response(200);
+  } else {
+    response = new Response(500);
+  }
+
+  return response;
 }
